lcd: WIP - merge st7920 and hd44780 code into one module
