+++ config/example-extras.cfg
@@ -326,8 +326,9 @@
 #pins = ^!ar31, ^!ar33, ^!ar35, ^!ar41
 
 
-# Reprap "2004" display
-#[lcd_hd44780]
+# Reprap "20x04" display
+#[lcd]
+#lcd_type: hd44780
 #rs_pin: ar16
 #e_pin: ar17
 #d4_pin: ar23
@@ -335,9 +336,9 @@
 #d6_pin: ar27
 #d7_pin: ar29
 
-
-# LCD ST7920
-#[lcd_st7920]
-#cs_pin:
-#sclk_pin:
-#sid_pin:
+# Reprap "128x64" graphics display
+#[lcd]
+#lcd_type: st7920
+#cs_pin: ar52
+#sclk_pin: ar50
+#sid_pin: ar51
